+++ src/lab06/Plataforma.java
@@ -4,7 +4,7 @@ import java.util.HashSet;
 
 public class Plataforma extends Jogo {
 
-	public Plataforma(String nome, int valor, HashSet<Jogabilidade> jogabilidades) throws Exception {
+	public Plataforma(String nome, double valor, HashSet<Jogabilidade> jogabilidades) throws Exception {
 		super(nome, valor, jogabilidades);
 	}
